Move tests to test directory.
